+++ front/src/views/RegularPage/RegularPageContent.ts
@@ -1,0 +1,4 @@
+export interface RegularPageContent {
+  izenburua: string;
+  edukia: string;
+}
