+++ front/src/views/RegularPage/index.ts
@@ -1,0 +1,1 @@
+export { RegularPage } from './RegularPage';
